+++ packages/web/src/components/DiffStack.vue
@@ -250,6 +250,7 @@ import {
 } from '../composables/useStackScroll';
 import DiffView from './DiffView.vue';
 import ViewFileButton from './ViewFileButton.vue';
+import CopyPathButton from './CopyPathButton.vue';
 import WrapToggle from './WrapToggle.vue';
 
 const props = defineProps<{
@@ -1296,6 +1297,7 @@ defineExpose({
         <span class="letter mono" :data-status="item.status">{{ statusLetter(item.status) }}</span>
         <span class="path mono" :title="item.path">{{ item.path }}</span>
         <ViewFileButton :path="item.path" />
+        <CopyPathButton :path="item.path" />
         <span v-if="item.uncommitted" class="uncommitted-tag mono">[uncommitted]</span>
         <span class="stats mono">
           <span v-if="item.stats.insertions" class="count-add">+{{ item.stats.insertions }}</span>